+++ src/lightsclient/MyReader.java
@@ -77,11 +77,11 @@ public class MyReader {
 				// m or p
 				if (fileName.equals("m")) {
 					// get times
-					ArrayList<Double> m = new ArrayList<Double>();
+					ArrayList<Float> m = new ArrayList<Float>();
 					for (String mLine : lines) {
 						if (!mLine.equals("")) {
 							String[] elements = mLine.split(" ");
-							m.add(Double.parseDouble(elements[0]));
+							m.add(Float.parseFloat(elements[0]));
 						}
 					}
 
@@ -91,11 +91,11 @@ public class MyReader {
 					// p
 
 					// get times
-					ArrayList<Double> p = new ArrayList<Double>();
+					ArrayList<Float> p = new ArrayList<Float>();
 					for (String pLine : lines) {
 						if (!pLine.equals("")) {
 							String[] elements = pLine.split(" ");
-							p.add(Double.parseDouble(elements[0]));
+							p.add(Float.parseFloat(elements[0]));
 						}
 					}
