+++ src/lightsclient/Phrase.java
@@ -5,22 +5,22 @@ public class Phrase {
 		int[] temp = new int[events.length];
 
 		// get average time between events
-		double averageTime = 0;
-		double t1 = events[0].getTime();
+		float averageTime = 0;
+		float t1 = events[0].getTime();
 		for (int i = 1; i < events.length; i++) {
-			double t2 = events[i].getTime();
+			float t2 = events[i].getTime();
 			averageTime += t2 - t1;
 			t1 = t2;
 		}
-		averageTime = averageTime / (double) events.length;
-//		System.out.println(averageTime);
+		averageTime = averageTime / (float) events.length;
+		// System.out.println(averageTime);
 
 		// attempt splitting up by time between notes
 		// boolean timeSplit = false;
 		int startIndex = 0;
 		int currentPhrase = 0;
 		for (int i = 1; i < events.length; i++) {
-			double timeDiff = events[i].getTime() - events[i - 1].getTime();
+			float timeDiff = events[i].getTime() - events[i - 1].getTime();
 			// System.out.println(events[i].getTime() + "-" + events[i -
 			// 1].getTime() + "=" + timeDiff);
 			if (timeDiff > averageTime) {
@@ -45,7 +45,7 @@ public class Phrase {
 
 	public static boolean similar(Event[] phrases, Event[] recentEvents) {
 		int numBad = 0;
-		
+
 		System.out.println("phrases: " + phrases.length + ", recentEvents: " + recentEvents.length);
 
 		for (int i = 0; i < recentEvents.length; i++) {
